etl/requirements.txt
psycopg2==2.9.9
flake8==4.0
python-dotenv==1.0.0
elasticsearch==8.10.1
backoff==2.2.1